Can now delete line items Better validation on delete

+++ admin/delete-line.php
@@ -16,7 +16,7 @@
         die("invalid line id"); // TODO: better error
     }
 
-    if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    if ($_SERVER['REQUEST_METHOD'] === 'POST' && $lineID == $_POST["line-id"]) {
         $query = 'DELETE FROM HomeInfoLines WHERE HomeInfoLineID = ?';
         $lineStmt = $pdo->prepare($query);
         $lineStmt->execute([$lineID]);
@@ -32,6 +32,7 @@
 <div id="delete-line">
     <h4> Are you sure you want to delete this line of text?</h4>
     <form method="post">
+        <input type="hidden" name="line-id" value="<?= $lineID ?>"/>
         <button class="btn waves-effect waves-light submit red white-text" type="submit" name="action">Delete Line</button>
     </form>
 </div>
