+++ admin/delete-user.php
@@ -1,18 +1,20 @@
 <?php
     require_once(dirname(__FILE__)."/init-admin.php");
 
+    $userID = $_GET["id"];
+
     $query = 'SELECT FirstName, LastName FROM Users WHERE UserID = ?';
     $stmt = $pdo->prepare($query);
-    $stmt->execute([$_GET["id"]]);
+    $stmt->execute([$userID]);
     $user = $stmt->fetch();
     if ($user == NULL) {
         die("invalid user id"); // TODO: better error
     }
 
-    if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    if ($_SERVER['REQUEST_METHOD'] === 'POST' && $userID == $_POST["user-id"]) {
         $query = 'DELETE FROM Users WHERE UserID = ?';
         $stmt = $pdo->prepare($query);
-        $stmt->execute([$_GET["id"]]);
+        $stmt->execute([$userID]);
         header("Location: view-users.php");
     }
 
@@ -25,7 +27,7 @@
 <div id="delete-user">
     <h4> Are you sure you want to delete <?= $user["FirstName"] ?> <?= $user["LastName"] ?>? </h4>
     <form method="post">
-        <input type="hidden" name="user-id" value="<?= $_GET['id'] ?>"/>
+        <input type="hidden" name="user-id" value="<?= $userID ?>"/>
         <button class="btn waves-effect waves-light submit red white-text" type="submit" name="action">Delete User</button>
     </form>
 </div>
